+++ imod/logging/__init__.py
@@ -60,7 +60,7 @@ If you want to integrate imod-python logging into your own logging framework
 
 from imod.logging._loggerholder import _LoggerHolder
 from imod.logging.config import LoggerType, configure
-from imod.logging.ilogger import ILogger , log_with_level # noqa: I001
+from imod.logging.ilogger import ILogger  # noqa: I001
 from imod.logging.loglevel import LogLevel
 
 logger = _LoggerHolder()
